CA-679 Add a node pool service account and a variable to give it GCR read access (#11)
Sam's image is currently stored on broad-dsp-gcr-public. We need to give our node pool service account object reader privileges on that project in order for it to pull images from there.

This PR adds an explicitly created node pool service account instead of using the default service account. It also adds a variable to terra-cluster to allow specifying which projects to give storage object viewer read priveleges to.